Initial

+++ src/main/java/org/jahdoo/common/event/ServerEvents.java
@@ -3,7 +3,6 @@ package org.jahdoo.common.event;
 import net.casual.arcade.dimensions.level.CustomLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.sounds.SoundEvents;
-import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.projectile.Arrow;
 import net.minecraft.world.phys.BlockHitResult;
@@ -19,7 +18,6 @@ import net.neoforged.neoforge.event.entity.player.PlayerInteractEvent;
 import net.neoforged.neoforge.event.entity.player.UseItemOnBlockEvent;
 import net.neoforged.neoforge.event.tick.LevelTickEvent;
 import net.neoforged.neoforge.event.tick.PlayerTickEvent;
-import net.neoforged.neoforge.network.PacketDistributor;
 import org.jahdoo.JahdooMod;
 import org.jahdoo.ascension.ability.abilities.dimensional_recall.DimensionalRecall;
 import org.jahdoo.ascension.ability.abilities.nova_smash.NovaSmash;
@@ -31,15 +29,13 @@ import org.jahdoo.ascension.attachments.player_abilities.TripleJump;
 import org.jahdoo.common.entities.ITamableEntity;
 import org.jahdoo.common.networking.server2client.InstanceSyncS2CP;
 import org.jahdoo.common.networking.server2client.WalletSyncS2CP;
-import org.jahdoo.common.registers.AttachmentReg;
 import top.theillusivec4.curios.api.event.CurioAttributeModifierEvent;
 
-import static net.minecraft.sounds.SoundSource.*;
-import static net.neoforged.neoforge.network.PacketDistributor.*;
+import static net.minecraft.sounds.SoundSource.PLAYERS;
+import static net.neoforged.neoforge.network.PacketDistributor.sendToPlayer;
 import static org.jahdoo.common.event.event_helpers.CopyPasteEvent.copyPasteBlockProperties;
 import static org.jahdoo.common.event.event_helpers.EventHelpers.*;
 import static org.jahdoo.common.registers.AttachmentReg.*;
-import static org.jahdoo.common.registers.AttachmentReg.SAVE_DATA;
 
 
 @EventBusSubscriber(modid = JahdooMod.MOD_ID)
@@ -119,27 +115,30 @@ public class ServerEvents {
             if(!cLevel.hasData(INSTANCE_DATA)) return;
             var data = cLevel.getData(INSTANCE_DATA);
 
-            data.incrementTicks();
+            var difficulty = data.getDifficulty();
 
-            for (var player : cLevel.players()) {
-                sendToPlayer(player, new InstanceSyncS2CP(data));
-                var remaining = data.getMaxTime() - data.getTicks();
-                var lessThan20Seconds = remaining <= 400;
-                var lessThan10Seconds = remaining <= 200;
+            if(!difficulty.isEmpty()){
+                data.incrementTicks();
+                for (var player : cLevel.players()) {
+                    sendToPlayer(player, new InstanceSyncS2CP(data));
+                    var remaining = data.getMaxTime() - data.getTicks();
+                    var lessThan20Seconds = remaining <= 400;
+                    var lessThan10Seconds = remaining <= 200;
 
-                if(lessThan20Seconds && (data.getTicks() % 20) == 0){
-                    var pitch = (float) Math.abs((remaining / 10) - 38) / 19;
-                    player.playNotifySound(SoundEvents.NOTE_BLOCK_BASS.value(), PLAYERS, 1, pitch);
-                    player.playNotifySound(SoundEvents.WARDEN_HEARTBEAT, PLAYERS, 1, 0.8F);
-                }
+                    if (lessThan20Seconds && (data.getTicks() % 20) == 0) {
+                        var pitch = (float) Math.abs((remaining / 10) - 38) / 19;
+                        player.playNotifySound(SoundEvents.NOTE_BLOCK_BASS.value(), PLAYERS, 1, pitch);
+                        player.playNotifySound(SoundEvents.WARDEN_HEARTBEAT, PLAYERS, 1, 0.8F);
+                    }
 
-                if(lessThan20Seconds && (data.getTicks() % (lessThan10Seconds ? 10 : 20)) == 0){
-                    player.playNotifySound(SoundEvents.WARDEN_HEARTBEAT, PLAYERS, 1, 0.8F);
-                }
+                    if (lessThan20Seconds && (data.getTicks() % (lessThan10Seconds ? 10 : 20)) == 0) {
+                        player.playNotifySound(SoundEvents.WARDEN_HEARTBEAT, PLAYERS, 1, 0.8F);
+                    }
 
-                if(remaining == 0) {
-                    player.playNotifySound(SoundEvents.ALLAY_DEATH, PLAYERS, 1, 0.8F);
-                    player.kill();
+                    if (remaining == 0) {
+                        player.playNotifySound(SoundEvents.ALLAY_DEATH, PLAYERS, 1, 0.8F);
+                        player.kill();
+                    }
                 }
             }
         }
